refactor(workspace): 优化工作空间及文件夹相关数据结构和逻辑
- 删除workspace的key字段，改用id作为唯一标识符
- 修改相关数据库mapper，将workspace_key替换为workspace_id
- 更新Folder实体及相关方法，使用workspace_id替代workspace_key
- 修改Folder相关数据库操作，支持workspace_id条件
- 调整Folder handler代码，支持使用workspace_id参数
- 引入工作空间缓存，简化工作空间切换流程，提升性能
- 更新认证中间件，从缓存获取工作空间ID
- 修改用户相关代码，优化结构与异常处理
- 细化错误码，新增工作空间删除限制错误码
- 更新前端上下文结构，使用workspace_id
- 重构工作空间创建、更新、删除及切换接口
- 代码格式及模块导入调整，整理imports和模块声明
- 增加Tag模块路由支持，并注册至主路由
- 修正序列号处理SQL语句，确保工作空间范围正确
- 修改错误处理逻辑，支持带参数错误显示
- 修改项目中id类型统一为u64，增强类型一致性和安全性
- 移除多余代码与序号字段，简化代码逻辑和维护成本
- 更新国际化信息，新增工作空间相关提示语

+++ uvwa-rs/src/models/folder.rs
@@ -8,9 +8,9 @@ use validator::Validate;
 #[serde(rename_all = "camelCase")]
 pub struct FolderResp {
     #[serde(serialize_with = "to_str")]
-    pub id: i32,
+    pub id: u64,
     #[serde(serialize_with = "to_str")]
-    pub parent_id: i32,
+    pub parent_id: u64,
     pub name: String,
     pub description: Option<String>,
     pub seq: i32,
@@ -34,7 +34,7 @@ impl From<Folder> for FolderResp {
 #[serde(rename_all = "camelCase")]
 pub struct CreateFolderReq {
     #[serde(deserialize_with = "to_number")]
-    pub parent_id: i32,
+    pub parent_id: u64,
     #[validate(length(min = 1))]
     pub name: String,
     pub description: Option<String>,
@@ -52,6 +52,6 @@ pub struct UpdateFolderReq {
 #[serde(rename_all = "camelCase")]
 pub struct MoveFolderReq {
     #[serde(deserialize_with = "to_number")]
-    pub parent_id: i32,
+    pub parent_id: u64,
     pub seq: i32,
 }
